refactor(agent): extract auth mode types from main (#216)
